# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FacilConsulta API",
        "description": "Documentação oficial da FacilConsulta API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor Local"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/doctors": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Listar médicos",
                "operationId": "1ff2acc28425ccb6474cd7925c28ddd9",
                "parameters": [
                    {
                        "name": "specialty",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de médicos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "doctors": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Doctor"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/doctors/available": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Listar médicos disponíveis por data",
                "operationId": "6f344c7cedaf25c6d7005d972a24073a",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "specialty",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Quantidade de dias consecutivos a partir da data inicial",
                        "schema": {
                            "type": "integer",
                            "maximum": 14,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Médicos disponíveis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "doctors": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Doctor"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Doctor": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "document": {
                        "type": "string"
                    },
                    "crm": {
                        "type": "string"
                    },
                    "specialty": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "experience_years": {
                        "type": "integer"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "schedules": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DoctorSchedule"
                        }
                    },
                    "availability": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "date": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "weekday": {
                                    "type": "integer"
                                },
                                "slots": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "13:00"
                                    }
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "DoctorSchedule": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "weekday": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_time": {
                        "type": "string",
                        "example": "09:00"
                    },
                    "end_time": {
                        "type": "string",
                        "example": "12:00"
                    },
                    "slot_duration": {
                        "type": "integer",
                        "example": 30
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}